Еще исправления по замечаниям
